% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read data of a given year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years}
}
\value{
A file with the correspondent data of those years.
}
\description{
When given some years this function will try to read the data of those years
and if there is a not valid year it will throw a warning.
}
\examples{
\dontrun{
my_data <- fars_read_years(1999:2003)}

}
