% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the values}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list with all the year to examine}
}
\value{
A table with years as rows and months as columns
}
\description{
Just a function which will create a table with the years as the rows and the
number of values for each month as the new value for each pair year/month
}
\examples{
\dontrun{
again_data<-fars_summarize_years(2010:2015)
}

}
