% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string character or number with the wanted year of the data}
}
\value{
The filepath correspondent to the data of that year.
}
\description{
What this function does is to add the year to a filepath so that it now
refers to the data of that year and print the new filepath.
}
\examples{
make_filename("2017")
make_filename(1999)

}
