% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_script.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Ploting map
This function uses names, generated by function make_filename to (in case of existense)
 import .csv files  by fars_read funnction.
With the specified state number all accidents are plotted on the map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a number of a state in US as integer}

\item{year}{an input year as integer}
}
\value{
this function returns a map with total number of accidents in the input state number
}
\description{
Ploting map
This function uses names, generated by function make_filename to (in case of existense)
 import .csv files  by fars_read funnction.
With the specified state number all accidents are plotted on the map
}
\examples{
\dontrun{
   fars_map_state(45, 2015)
   fars_read_years(2016) #example of warning (return error!)
   fars_read_years(60, 2016) #example of warning (return error!)
}

}
