% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_script.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Only year based csv file loading
This file uses the only value of a year to import csv files having names, generated by \code{\link{make_filename}}
function. Then it  selects year and month from each imported files and store it an a list.}
\usage{
fars_read_years(years)
}
\value{
List of dataframes, where each data frame corresponds to initial .csv file and have two columns:
month and year is returned. In case of absence of the requested .csv file to import, list value for it is NULL
and warning message is produced.
}
\description{
Only year based csv file loading
This file uses the only value of a year to import csv files having names, generated by \code{\link{make_filename}}
function. Then it  selects year and month from each imported files and store it an a list.
}
\examples{
fars_read_years(years=c(2011,2012,2013))
fars_read_years(years=seq(from=2011, to=2015))

}
