% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_script.R
\name{make_filename}
\alias{make_filename}
\title{File Naming
This function make character name using existing pattern and year, so final name for year=2012
becomes "accident_2012.csv.bz2 "}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A parameter which identifies the value of year, used in built file name}
}
\value{
This function returns object of class "character" with the constructed file name
}
\description{
File Naming
This function make character name using existing pattern and year, so final name for year=2012
becomes "accident_2012.csv.bz2 "
}
\examples{
\dontrun{filename<-make_filename(year=2012)}

}
